+++ src/vue/forms/WithdrawalForm.vue
@@ -2,6 +2,16 @@
   <div class="withdrawal">
     <template v-if="isLoaded">
       <template v-if="assets.length">
+        <button
+          v-ripple
+          type="button"
+          class="app__button-raised withdrawal__form-fiat-btn"
+          :disabled="formMixin.isDisabled"
+          form="withdrawal-form"
+          @click="fiatWithdrawalFormShown = true"
+        >
+          {{ 'withdrawal-form.withdrawal-fiat' | globalize }}
+        </button>
         <form
           @submit.prevent="isFormValid() && showConfirmation()"
           id="withdrawal-form"
@@ -198,25 +208,23 @@ export default {
     Loader,
   },
   mixins: [FormMixin],
-  data () {
-    return {
-      isLoaded: false,
-      isFailed: false,
-      form: {
-        asset: {},
-        amount: '',
-        address: '',
-      },
-      assets: [],
-      MIN_AMOUNT: config.MIN_AMOUNT,
-      fixedFee: EMPTY_FEE,
-      percentFee: EMPTY_FEE,
-      feesDebouncedRequest: null,
-      isFeesLoadPending: false,
-      isFeesLoadFailed: false,
-      DECIMAL_POINTS: config.DECIMAL_POINTS,
-    }
-  },
+  data: () => ({
+    isLoaded: false,
+    isFailed: false,
+    form: {
+      asset: {},
+      amount: '',
+      address: '',
+    },
+    assets: [],
+    MIN_AMOUNT: config.MIN_AMOUNT,
+    fixedFee: EMPTY_FEE,
+    percentFee: EMPTY_FEE,
+    feesDebouncedRequest: null,
+    isFeesLoadPending: false,
+    isFeesLoadFailed: false,
+    DECIMAL_POINTS: config.DECIMAL_POINTS,
+  }),
   validations () {
     return {
       form: {
@@ -234,6 +242,7 @@ export default {
   },
   computed: {
     ...mapGetters({
+      wallet: vuexTypes.wallet,
       accountId: vuexTypes.accountId,
       balances: vuexTypes.accountBalances,
     }),
@@ -398,5 +407,9 @@ export default {
 .withdrawal__table-description {
   opacity: 0.6;
   font-size: 1.2rem;
+}
+
+.withdrawal__form-fiat-btn {
+  margin-bottom: 2.4rem;
 }
 </style>