+++ src/vuex/index.js
@@ -9,6 +9,8 @@ import keyValue from './key-value.module'
 
 import { movementsHistoryModule } from '@/vue/modules/movements-history/store'
 import { feesModule } from '@/vue/modules/fees/store'
+import { withdrawalFiatModule } from '@/vue/modules/withdrawal-fiat/store'
+import { withdrawalFiatBankModule } from '@/vue/modules/withdrawal-fiat-bank/store'
 
 import { vuexTypes } from '@/vuex/types'
 import { sessionStoragePlugin } from './plugins/session-storage'
@@ -43,6 +45,8 @@ const store = new Vuex.Store({
     // namespaced local modules (used by a specific set of components)
     'movements-history': movementsHistoryModule,
     'fees': feesModule,
+    'withdrawal-fiat': withdrawalFiatModule,
+    'withdrawal-fiat-bank': withdrawalFiatBankModule,
   },
   plugins: [sessionStoragePlugin],
 })