+++ CHANGELOG.md
@@ -11,7 +11,8 @@ for further information about branching and tagging conventions.
 ### Added
 - Fees module
 - User-friendly data displaying in the input fields. If the data does not fit in a field, then three dots are added at the end
-- Displaying TokenD logo & load spinner while loading the page 
+- Displaying TokenD logo & load spinner while loading the page
+- Withdrawal fiat by the bank information form module
 
 ### Changed
 - Now using new account endpoint for loading the fees